+++ in_app_review/lib/in_app_review.dart
@@ -1,17 +1,11 @@
 import 'dart:async';
-import 'dart:io';
 
-import 'package:flutter/services.dart';
-import 'package:package_info/package_info.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:in_app_review_platform_interface/in_app_review_platform_interface.dart';
 
 class InAppReview {
   InAppReview._();
 
   static final InAppReview instance = InAppReview._();
-
-  static const MethodChannel _channel =
-      const MethodChannel('dev.britannio.in_app_review');
 
   /// Check's if the device is able to show a review dialog.
   ///
@@ -21,7 +15,7 @@ class InAppReview {
   /// IOS devices must be running **IOS version 10.3** or higher.
   ///
   /// MacOS devices must be running **MacOS version 10.14** or higher
-  Future<bool> isAvailable() => _channel.invokeMethod('isAvailable');
+  Future<bool> isAvailable() => InAppReviewPlatform.instance.isAvailable();
 
   /// Attempts to show the review dialog. It's recommended to first check if
   /// this cannot be done via [isAvailable]. If it is not available then
@@ -35,45 +29,22 @@ class InAppReview {
   /// https://developer.android.com/guide/playcore/in-app-review#when-to-request
   /// https://developer.apple.com/design/human-interface-guidelines/ios/system-capabilities/ratings-and-reviews/
   /// https://developer.apple.com/design/human-interface-guidelines/macos/system-capabilities/ratings-and-reviews/
-  Future<void> requestReview() => _channel.invokeMethod('requestReview');
+  Future<void> requestReview() => InAppReviewPlatform.instance.requestReview();
 
-  /// Opens the Play Store on Android and opens the App Store with a review
-  /// screen on IOS & MacOS. [appStoreId] is required for IOS & MacOS.
+  /// Opens the Play Store on Android, the App Store with a review
+  /// screen on IOS & MacOS and the Microsoft Store on Windows.
+  ///
+  /// [appStoreId] is required for IOS & MacOS.
+  /// [windowsStoreId] is required for Windows.
   Future<void> openStoreListing({
-    /// Required for IOS & MacOS
+    /// Required for IOS & MacOS.
     String appStoreId,
-  }) async {
-    final bool isIOS = Platform.isIOS;
-    final bool isMacOS = Platform.isMacOS;
-    final bool isAndroid = Platform.isAndroid;
 
-    if (isIOS || isMacOS) {
-      assert(appStoreId != null);
-
-      final Uri uri = Uri(
-        scheme: isIOS ? 'https' : 'macappstore',
-        host: 'apps.apple.com',
-        path: 'app/id$appStoreId',
-        queryParameters: {'action': 'write-review'},
+    /// Required for Windows.
+    String windowsStoreId,
+  }) =>
+      InAppReviewPlatform.instance.openStoreListing(
+        appStoreId: appStoreId,
+        windowsStoreId: windowsStoreId,
       );
-
-      await _launchUrl(uri.toString());
-    } else if (isAndroid) {
-      final PackageInfo packageInfo = await PackageInfo.fromPlatform();
-
-      final String packageName = packageInfo.packageName;
-
-      await _launchUrl(
-        'https://play.google.com/store/apps/details?id=$packageName',
-      );
-    } else {
-      throw UnsupportedError('Platform not supported');
-    }
-  }
-
-  Future<void> _launchUrl(String url) async {
-    if (await canLaunch(url)) {
-      await launch(url);
-    }
-  }
 }